Allow reference areas which extend past the canvas bounds
I’m working on a chart with several reference areas, some of which extend past the canvas bounds. Though I want to draw the part of the reference area that lies on the canvas, it doesn’t make sense to draw the part that spills out.

This adds a clip path to the chart which it applies to all reference areas. It’s applied using a presentation attribute, so it’s easy to override if need be.

It relaxes the constraint on reference areas. Any reference area which is partly in bounds will draw.